requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
chromadb>=0.4.0
ollama>=0.0.11
streamlit>=1.28.0
pypdf>=3.17.0
python-docx>=0.8.11
requests>=2.31.0